Fix bug

+++ backend/english_learning_app/flashcards/managers/bases.py
@@ -3,4 +3,5 @@ from django.db import models
 
 class BaseManager(models.Manager):
     def bulk_delete(self, ids):
-        self.get_queryset().filter(pk__in=ids).delete()
+        if isinstance(ids, list) and len(ids) > 0:
+            self.get_queryset().filter(pk__in=ids).delete()
